requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
streamlit==1.44.1
requests==2.32.3
pillow==11.2.1
pydantic==2.11.3
pydantic-settings==2.8.1
chromadb==0.6.3
google-generativeai==0.8.5
